+++ src/rosecommon/src/packets/cli_createcharreq.cpp
@@ -16,7 +16,7 @@ CliCreateCharReq::CliCreateCharReq(uint8_t buffer[MAX_PACKET_SIZE]) : CRosePacke
 	*this >> name_;
 }
 
-CliCreateCharReq::CliCreateCharReq(uint8_t race, uint8_t stone, uint8_t hair, uint8_t face, uint8_t weapon, uint8_t zone, const std::string &name) : CRosePacket(ePacketType::PAKCS_CREATE_CHAR_REQ), race_(race), stone_(stone), hair_(hair), face_(face), weapon_(weapon), zone_(zone), name_(name) {}
+CliCreateCharReq::CliCreateCharReq(uint8_t race, uint8_t stone, uint8_t hair, uint8_t face, uint8_t weapon, uint16_t zone, const std::string &name) : CRosePacket(ePacketType::PAKCS_CREATE_CHAR_REQ), race_(race), stone_(stone), hair_(hair), face_(face), weapon_(weapon), zone_(zone), name_(name) {}
 
 uint8_t CliCreateCharReq::race() const {
 	return race_;
@@ -38,7 +38,7 @@ uint8_t CliCreateCharReq::weapon() const {
 	return weapon_;
 }
 
-uint8_t CliCreateCharReq::zone() const {
+uint16_t CliCreateCharReq::zone() const {
 	return zone_;
 }
